Wed, Dec 18, 2024, 2:50 PM -06:00

+++ fuel/fuel.py
@@ -48,7 +48,7 @@ def is_percentage(s):
 def divide(a, b):
     if b == 0:
         raise ZeroDivisionError
-    return round(int(a/b),2)
+    return round(float(a/b),2)
 
 def is_int(num):
     try:
